Character tables consolidated
