+++ src/App.js
@@ -13,7 +13,6 @@ import Signup from './pages/Signup'
 
 function App() {
 	const { user } = useAuthContext()
-	console.log(user)
 
 	return (
 		<div className="App">
