+++ src/Components/Character/DTAForm.js
@@ -7,7 +7,7 @@ import Dialog from '@mui/material/Dialog'
 import DialogActions from '@mui/material/DialogActions'
 import DialogContent from '@mui/material/DialogContent'
 import DialogTitle from '@mui/material/DialogTitle'
-import { handleAttributeUpdate, createEmptyTableData } from '../../Utils'
+import { handleAttributeUpdate, createEmptyTableData } from '../Utils'
 
 const DTAForm = (props) => {
 	const {
