+++ src/Components/Character/DTARow.js
@@ -3,7 +3,7 @@ import TableCell from '@mui/material/TableCell'
 import TableRow from '@mui/material/TableRow'
 import IconButton from '@mui/material/IconButton'
 import EditIcon from '@mui/icons-material/Edit'
-import { dieTestArthaCells, skillsLearningCells } from '../../TableConfig'
+import { dieTestArthaCells, skillsLearningCells } from '../TableConfig'
 
 const DTARow = (props) => {
 	const { attribute, row, handleDialogToggle, setDialogData, setDialogType } =
